// src/features/transport-services/api/transport-sheets-types.ts
export type TransportSheetFormData = {
  sheetNumber?: number;
  organizationId: string;
  reportDate: string;
  destinationName: string;
  destinationAddress: string;
  destinationCity: string;
  destinationProvince: string;
  destinationNotes: string;
  clientFiscalCode: string;
  clientFirstName: string;
  clientLastName: string;
  clientPhone: string;
  clientAddress: string;
  clientCity: string;
  clientProvince: string;
  clientAslNumber: string;
  clientAslMunicipality: string;
  clientNotes: string;
  vehiclePlate: string;
  startTime: string;
  endTime: string;
  kmDeparture: string;
  kmArrival: string;
  notes: string;
  volunteerIds: string[];
  volunteerDisplayNames?: string[];
};